Moved (Backend/Frontend)Init files to `App\Component\Application`
Bye bye `Common` namespace… Sorry to see you go… Not. Time for a beer,
said Chuck Norris.
