+++ src/Component/Application/FrontendInit.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Frontend;
+namespace App\Component\Application;
 
 /**
  * This class will initiate the frontend-application
  */
-class Init extends \Common\Core\Init
+class FrontendInit extends Init
 {
     protected $allowedTypes = ['Frontend', 'FrontendAjax'];
