+++ src/Controller/ForkController.php
@@ -6,10 +6,10 @@ use App\Component\Application\ApplicationInterface;
 use App\Exception\RedirectException;
 use Backend\Core\Engine\Ajax as BackendAjax;
 use Backend\Core\Engine\Backend;
-use Backend\Init as BackendInit;
+use App\Component\Application\BackendInit ;
 use Frontend\Core\Engine\Ajax as FrontendAjax;
 use Frontend\Core\Engine\Frontend;
-use Frontend\Init as FrontendInit;
+use App\Component\Application\FrontendInit ;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Response;
 use Twig_Error;
